+++ src/components/solicitudes/SolicitudCard.tsx
@@ -4,16 +4,16 @@ import Link from 'next/link';
 import { useSession } from '@/context/SessionProvider';
 import { HelpRequestAdditionalInfo, HelpRequestData, SelectedHelpData } from '@/types/Requests';
 import AsignarSolicitudButton from '@/components/AsignarSolicitudButton';
-import SolicitudHelpCount from '@/components/SolicitudHelpCount';
+import SolicitudHelpCount from '@/components/solicitudes/SolicitudHelpCount';
 import PhoneInfo from '@/components/PhoneInfo';
-import DeleteHelpRequest from './DeleteHelpRequest';
+import DeleteHelpRequest from '../DeleteHelpRequest';
 import { textWithEllipsis } from '@/helpers/utils';
 import { useTowns } from '@/context/TownProvider';
 import { useRole } from '@/context/RoleProvider';
 import { useState } from 'react';
-import ChangeUrgencyHelpRequest from './ChangeUrgencyHelpRequest';
-import ChangeStatusButton from './ChangeStatusButton';
-import ChangeCRMStatus from './ChangeCRMStatus';
+import ChangeUrgencyHelpRequest from '../ChangeUrgencyHelpRequest';
+import ChangeStatusButton from '../ChangeStatusButton';
+import ChangeCRMStatus from '../ChangeCRMStatus';
 import { UserRoles } from '@/helpers/constants';
 import CRMNotes from '@/components/CRMNotes';
